Remove old implementations to keep only the working one
